Aggiunta gestione location

+++ src/main/java/it/ariannamondo/mag/rest/installazione/StatiInstallazione.java
@@ -11,19 +11,19 @@ package it.ariannamondo.mag.rest.installazione;
  * //stato del task 0=ATTESA,1=AVVIATO,2=CONCLUSO,3=ANNULLATO,4=SOSPESO,5=ELIMINATA/NON ATTIVA
  */
 public enum StatiInstallazione {
-    ATTESA(0),
-    AVVIATA(1),
-    CONLCUSA(2),
-    ANNULLATA(3),
-    SOSPESA(4),
-    ELIMINATA(5);
+    ATTESA((short)0),
+    AVVIATA((short)1),
+    CONLCUSA((short)2),
+    ANNULLATA((short)3),
+    SOSPESA((short)4),
+    ELIMINATA((short)5);
     
-    private final int stato;
-    private StatiInstallazione(int stato){
+    private final short stato;
+    private StatiInstallazione(short stato){
         this.stato = stato;
     }
 
-    public int getStato() {
+    public short getStato() {
         return stato;
     }
